InstallerPage component-server
